Refactoring to add some structure
