+++ pkg/downstream/types/types.go
@@ -1,0 +1,6 @@
+package types
+
+type Downstream struct {
+	ClusterID string
+	Name      string
+}
