+++ pkg/registry/types/types.go
@@ -1,0 +1,8 @@
+package types
+
+type RegistrySettings struct {
+	Hostname    string
+	Username    string
+	PasswordEnc string
+	Namespace   string
+}
