+++ pkg/task/task.go
@@ -1,4 +1,4 @@
-package app
+package task
 
 import (
 	"database/sql"
