+++ pkg/version/types/types.go
@@ -1,0 +1,7 @@
+package types
+
+type AppVersion struct {
+	Sequence     int64
+	UpdateCursor int
+	VersionLabel string
+}
